book_app_backend/requirements.txt
asgiref==3.9.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==3.4.2
distlib==0.3.4
dj-database-url==0.5.0
Django==5.2.4
django-cors-headers==3.11.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.7.0
gunicorn==20.1.0
idna==3.10
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.9.0
pytz==2022.1
requests==2.32.4
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.14.1
whitenoise==5.3.0
